feat(meal-data): Implement meal logging and UI enhancements - Added `handleSaveMealLog` mutation in `meal-data.tsx` to save meal logs via API. - Updated `MealRequest` and `NutritionResponse` in the Java backend to use Double instead of Integer for macronutrient values. - Added `@JsonCreator` to `MealType` enum in Java backend to correctly serialize meal types - Updated backend URL in `utils/index.ts`. - Implemented token check on `_layout.tsx` to redirect the user to login screen if token is expired.

+++ ameri/src/main/java/com/winnerx0/ameri/dto/request/MealRequest.java
@@ -56,16 +56,16 @@ public class MealRequest {
     public static class Macros {
 
         @NotNull(message = "Calories required")
-        private Integer calories;
+        private Double calories;
 
         @NotNull(message = "Carbs required")
-        private Integer carbs;
+        private Double carbs;
 
         @NotNull(message = "Protein required")
-        private Integer protein;
+        private Double protein;
 
         @NotNull(message = "Fat required")
-        private Integer fat;
+        private Double fat;
     }
 
 }
